refactor(bookings): extract checked-out helpers to break the availability/service cycle
`asset/availability.server.ts` imported `computeCheckedOutBreakdownForAsset` from
`booking/service.server.ts`, which imports `getAssetAvailability` back — a circular
import between two large server modules (pre-existing; surfaced in #2801 review).

Move `computeCheckedOutForAsset` + `computeCheckedOutBreakdownForAsset` (verbatim) into
a new `booking/checked-out.server.ts` that depends only on the pure
`checkout-attribution.ts` primitives and `@prisma/client`. `availability.server.ts` now
imports from there, so it no longer references `service.server.ts` at all and the cycle
is gone. Behavior is unchanged — the functions moved as-is, existing parity/availability
tests still cover them. Also retargets the four test files' imports/mocks and drops the
now-unused `CheckoutSession` import from `service.server.ts`.
